refactor: Entity 코드 리팩토링

+++ src/main/java/com/github/classpick/reservation/controller/dto/request/CreateReservationReq.java
@@ -19,7 +19,7 @@ public class CreateReservationReq {
     private String purpose;
 
     @NotNull
-    private Long people;
+    private Integer people;
 
     @NotNull
     private LocalDate date;
